remove train dependencies on tune
